patch: Check if build directory exists

+++ rollup.config.js
@@ -4,10 +4,14 @@ import fs from 'fs';
 
 const IS_PRODUCTION = !process.env.ROLLUP_WATCH;
 
-// Clear the build folder
-fs.rmSync("build/", {
-    recursive: true
-})
+// Clear the build folder if it exists
+if(fs.existsSync("build/")) {
+    fs.rmSync("build/", {
+        recursive: true
+    })
+} else {
+    fs.mkdirSync("build/");
+}
 
 // Copy root level stuff directly
 fs.cpSync("src/root", "build", { recursive: true });
